// types.ts
export declare const INTERFACE_MESSAGES: {
  readonly ENV_INFO: "ENV_INFO";
  readonly ON_SAFE_INFO: "ON_SAFE_INFO";
  readonly TRANSACTION_CONFIRMED: "TRANSACTION_CONFIRMED";
  readonly TRANSACTION_REJECTED: "TRANSACTION_REJECTED";
};
export type InterfaceMessageIds = keyof typeof INTERFACE_MESSAGES;

export declare type LowercaseNetworks =
  | "mainnet"
  | "morden"
  | "ropsten"
  | "rinkeby"
  | "goerli"
  | "kovan"
  | "xdai"
  | "energy_web_chain"
  | "volta"
  | "unknown";
export interface SafeInfo {
  safeAddress: string;
  network: LowercaseNetworks;
  ethBalance: string;
}
export interface InterfaceMessageToPayload {
  [INTERFACE_MESSAGES.ON_SAFE_INFO]: SafeInfo;
  [INTERFACE_MESSAGES.TRANSACTION_CONFIRMED]: {
    safeTxHash: string;
  };
  [INTERFACE_MESSAGES.ENV_INFO]: {
    txServiceUrl: string;
  };
  [INTERFACE_MESSAGES.TRANSACTION_REJECTED]: Record<string, unknown>;
}
export type InterfaceMessageProps<T extends InterfaceMessageIds> = {
  messageId: T;
  data: InterfaceMessageToPayload[T];
};
export declare type RequestId = number | string;

// Messaging
export enum Methods {
  sendTransactions = "sendTransactions",
  rpcCall = "rpcCall",
  getChainInfo = "getChainInfo",
  getSafeInfo = "getSafeInfo",
  getTxBySafeTxHash = "getTxBySafeTxHash",
  getSafeBalances = "getSafeBalances",
  signMessage = "signMessage",
  signTypedMessage = "signTypedMessage",
  getEnvironmentInfo = "getEnvironmentInfo",
  requestAddressBook = "requestAddressBook",
  wallet_getPermissions = "wallet_getPermissions",
  wallet_requestPermissions = "wallet_requestPermissions",
}
export declare type SDKRequestData<M extends Methods = Methods, P = unknown> = {
  id: RequestId;
  params: P;
  env: {
    sdkVersion: string;
  };
  method: M;
};
export declare type SDKMessageEvent = MessageEvent<SDKRequestData>;
export declare type SendTransactionsResponse = {
  safeTxHash: string;
};
// ...
export type _ChainInfo = {
  transactionService: string;
  chainId: string; // Restricted by what is returned by the CGW
  chainName: string;
  shortName: string;
  l2: boolean;
  description: string;
  rpcUri: RpcUri;
  safeAppsRpcUri: RpcUri;
  publicRpcUri: RpcUri;
  blockExplorerUriTemplate: BlockExplorerUriTemplate;
  nativeCurrency: NativeCurrency;
  theme: Theme;
  ensRegistryAddress?: string;
  gasPrice: GasPrice;
  disabledWallets: string[];
  features: FEATURES[];
};
export declare type ChainInfo = Pick<
  _ChainInfo,
  | "chainName"
  | "chainId"
  | "shortName"
  | "nativeCurrency"
  | "blockExplorerUriTemplate"
>;
// ...
export type TransactionDetails = {
  txId: string;
  executedAt?: number;
  txStatus: TransactionStatus;
  txInfo: TransactionInfo;
  txData?: TransactionData;
  detailedExecutionInfo?: DetailedExecutionInfo;
  txHash?: string;
  safeAppInfo?: SafeAppInfo;
};
export declare type GatewayTransactionDetails = TransactionDetails;
export type SafeBalanceResponse = {
  fiatTotal: string;
  items: Array<{
    tokenInfo: TokenInfo;
    balance: string;
    fiatBalance: string;
    fiatConversion: string;
  }>;
};
export declare type SafeBalances = SafeBalanceResponse;
export declare type EnvironmentInfo = {
  origin: string;
};
export declare type AddressBookItem = {
  address: string;
  chainId: string;
  name: string;
};
export declare type PermissionCaveat = {
  type: string;
  value?: unknown;
  name?: string;
};
export declare type Permission = {
  parentCapability: string;
  invoker: string;
  date?: number;
  caveats?: PermissionCaveat[];
};
export interface MethodToResponse {
  [Methods.sendTransactions]: SendTransactionsResponse;
  [Methods.rpcCall]: unknown;
  [Methods.getSafeInfo]: SafeInfo;
  [Methods.getChainInfo]: ChainInfo;
  [Methods.getTxBySafeTxHash]: GatewayTransactionDetails;
  [Methods.getSafeBalances]: SafeBalances[];
  [Methods.signMessage]: SendTransactionsResponse;
  [Methods.signTypedMessage]: SendTransactionsResponse;
  [Methods.getEnvironmentInfo]: EnvironmentInfo;
  [Methods.requestAddressBook]: AddressBookItem[];
  [Methods.wallet_getPermissions]: Permission[];
  [Methods.wallet_requestPermissions]: Permission[];
}
export declare type ErrorResponse = {
  id: RequestId;
  success: false;
  error: string;
  version?: string;
};
export type SuccessResponse<T = MethodToResponse[Methods]> = {
  id: RequestId;
  data: T;
  version?: string;
  success: true;
};
export declare const RPC_CALLS: {
  readonly eth_call: "eth_call";
  readonly eth_gasPrice: "eth_gasPrice";
  readonly eth_getLogs: "eth_getLogs";
  readonly eth_getBalance: "eth_getBalance";
  readonly eth_getCode: "eth_getCode";
  readonly eth_getBlockByHash: "eth_getBlockByHash";
  readonly eth_getBlockByNumber: "eth_getBlockByNumber";
  readonly eth_getStorageAt: "eth_getStorageAt";
  readonly eth_getTransactionByHash: "eth_getTransactionByHash";
  readonly eth_getTransactionReceipt: "eth_getTransactionReceipt";
  readonly eth_getTransactionCount: "eth_getTransactionCount";
  readonly eth_estimateGas: "eth_estimateGas";
};

export declare type RpcCallNames = keyof typeof RPC_CALLS;
export declare type RPCPayload<P = unknown[]> = {
  call: RpcCallNames;
  params: P | unknown[];
};
// ...
export declare type SignMessageParams = {
  message: string;
};
export interface TypedDataDomain {
  name?: string;
  version?: string;
  chainId?: BigNumberish;
  verifyingContract?: string;
  salt?: BytesLike;
}
export interface TypedDataTypes {
  name: string;
  type: string;
}
export declare type TypedMessageTypes = {
  [key: string]: TypedDataTypes[];
};
export declare type EIP712TypedData = {
  domain: TypedDataDomain;
  types: TypedMessageTypes;
  message: Record<string, any>;
};
export declare type SignTypedMessageParams = {
  typedData: EIP712TypedData;
};
export interface Transaction {
  to: string;
  value: string;
  data: string;
}